Update Model.m
The model class might benefit from differentiating the uncertainty associated with the initial states and the uncertainty associated with the model itself.  Therefore, this incorporates a new property for the initial states variance. If the initial states variances are not specified, they are set to be equal to the variance of process noise.